requirements.txt
python-telegram-bot==21.6
openai==1.55.0
python-dotenv==1.0.1
asyncpg==0.30.0
aiohttp==3.11.11
